pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[project]
name = "jidoushiwake"
version = "0.1.0"
description = "ScanSnap OCR to Yayoi CSV with review UI and learning"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.23",
    "SQLAlchemy>=2.0",
    "pydantic>=2.0",
    "python-multipart>=0.0.9",
    "PyQt6>=6.5",
    "psycopg2-binary>=2.9; platform_system == 'Windows'",
]

[project.optional-dependencies]
pdf = ["PyPDF2>=3.0", "pdfminer.six>=20231228"]
